billing-backend/payment-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
eventlet==0.33.3
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-json-logger==2.0.7
structlog==23.2.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
flower==2.0.1
bcrypt==4.1.2 